refactor(data): Data nach sdata.deprecated verschieben + uuid-Bug fixen
- sdata/data.py -> sdata/deprecated/data.py (DataFrame ersetzt Data). Data bleibt
  als sdata.Data nutzbar (Re-Export); Importeure in experiments/* angepasst;
  aus Coverage ausgeblendet.
- Vault-uuid-Bug: VaultSqliteIndex.update_from_metadata nutzt get_dict() statt
  get_sdict() (Data hat '!sdata_*'-Praefix, das get_sdict nicht erfasst) ->
  test_sqlitevault gruen (letzter Tiefen-Bug behoben).

Suite: 315 passed, 0 failed. Coverage 87%.
