+++ sdata/deprecated/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""Deprecated sdata-Module.
+
+Hier liegen abgelöste, nicht mehr weiterentwickelte Klassen. ``Data`` wird durch
+:class:`sdata.sclass.dataframe.DataFrame` ersetzt und nur aus
+Kompatibilitätsgründen weiter exportiert.
+
+Bewusst ohne ``from sdata import ...`` (Zirkularität) und aus der Coverage
+ausgeblendet.
+"""
